add: BestProduct 조회 repository 연결
- sum(quantity) 반환값이 long 이므로 totalSalesQuantity 에 대한 타입 변경
- 조회에 대한 반환값으로 원하는 필드만 조회하기 위해 projection 생성

+++ src/main/java/kr/hhplus/be/server/product/application/result/BestProductResult.java
@@ -3,6 +3,6 @@ package kr.hhplus.be.server.product.application.result;
 public record BestProductResult(
         long productId,
         String productName,
-        int totalSalesQuantity
+        long totalSalesQuantity
 ) {
 }
